+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - UCI understand debug command
 - UCI understand stop command
+- UCI understand go infinite command
 
 ### Added